More consistent Redux state clean up after code review
