+++ features/base/connection/actions.js
@@ -1,7 +1,10 @@
 import { conferenceInitialized } from '../conference';
 import JitsiMeetJS from '../lib-jitsi-meet';
 import { localParticipantJoined } from '../participants';
-import { createLocalTracks } from '../tracks';
+import {
+    createLocalTracks,
+    destroyLocalTracks
+} from '../tracks';
 
 import {
     CONNECTION_CREATED,
@@ -94,6 +97,35 @@ export function connectionInitialized(connection, room) {
             room,
             connection
         });
+    };
+}
+
+/**
+ * Leaves the conference, closes the connection and destroys local tracks.
+ *
+ * @returns {Function}
+ */
+export function destroy() {
+    return (dispatch, getState) => {
+        const state = getState();
+        const conference = state['features/base/conference'];
+        const connection = state['features/base/connection'];
+
+        let promise = Promise.resolve();
+
+        if (conference) {
+            promise = conference.leave();
+        }
+
+        if (connection) {
+            promise = promise
+                .then(() => connection.disconnect());
+        }
+
+        // XXX Local tracks might exist without conference and connection
+        // initialized, so we need to explicitly clean them here.
+        return promise
+            .then(() => dispatch(destroyLocalTracks()));
     };
 }
