+++ features/base/participants/reducer.js
@@ -1,3 +1,4 @@
+import { CONFERENCE_LEFT } from '../conference';
 import { ReducerRegistry } from '../redux';
 
 import {
@@ -125,6 +126,12 @@ function participant(state, action) {
  */
 ReducerRegistry.register('features/base/participants', (state = [], action) => {
     switch (action.type) {
+    /**
+     * Remove local participant when conference is left.
+     */
+    case CONFERENCE_LEFT:
+        return state.filter(p => !p.local);
+
     case PARTICIPANT_JOINED:
         return [ ...state, participant(undefined, action) ];
