+++ features/base/tracks/reducer.js
@@ -13,17 +13,14 @@ import {
 ReducerRegistry.register('features/base/tracks', (state = [], action) => {
     switch (action.type) {
     /**
-     * Remove participant's tracks after/when participant leaves.
+     * Remove remote participant's tracks after/when participant leaves.
+     * TODO: this can be removed when
+     * https://github.com/jitsi/lib-jitsi-meet/pull/174 will be merged.
      */
     case PARTICIPANT_LEFT:
-        return (
-            state.filter(
-                track => action.participant.local
-                    ? !track.isLocal()
-                    : (track.isLocal()
-                        || track.getParticipantId() !== action.participant.id)
-            )
-        );
+        return state.filter(
+            track => track.isLocal() ||
+                track.getParticipantId() !== action.participant.id);
 
     case TRACK_ADDED:
         return [...state, action.track];
